requirements.txt
aniso8601==9.0.1
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
flasgger==0.9.5
Flask==2.1.2
flask-restplus==0.13.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.1
MarkupSafe==2.1.1
mistune==2.0.3
numpy==1.23.0
pandas==1.4.3
Pillow==9.2.0
pycodestyle==2.8.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pyuwsgi==2.0.20
PyYAML==6.0
requests==2.28.1
six==1.16.0
toml==0.10.2
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.9
waitress==2.1.2
Werkzeug==2.1.2
zipp==3.8.0
